+++ tests/java/src/fr/ill/ics/cameo/test/TestBasicResponder.java
@@ -68,7 +68,7 @@ public class TestBasicResponder {
 				
 				// Start the application.
 				App responderApplication = server.start(applicationName, appArgs);
-				System.out.println("Started application " + responderApplication + " with state " + State.toString(responderApplication.getActualState()));
+				System.out.println("Started application " + responderApplication + " with state " + State.toString(responderApplication.getState()));
 
 				Requester requester = Requester.create(responderApplication, "responder");
 				
@@ -82,7 +82,7 @@ public class TestBasicResponder {
 				System.out.println("Created requester " + requester);
 			
 				// Check the state of the responder app.
-				System.out.println("Application " + responderApplication + " has state " + State.toString(responderApplication.getActualState()));
+				System.out.println("Application " + responderApplication + " has state " + State.toString(responderApplication.getState()));
 				
 				
 				// Send a simple message.